bring back spacing scale - sorry!
If the scale is 1.5 is still works to be divisible by 4’s or 8’s, so doesn’t leave us with half-pixels which was causing other issues for me with grid/gaps/etc.

The margin utility classes don’t have enough space for what I was needing them for, but I can deal with that later.