remove divider z-index inside group/cover
This is not worth fighting with. A divider should rarely have negative margin enough to go behind the content.

Startup Pro has it, and we added negative margin helper class to Reach Agency demo. But we should just not do worry about the overlap, or use less negative margin.

Any CSS i tried is too aggressive as a global/theme thing. We can do a one-off thing if needed.

Closes https://github.com/maithemewp/mai-engine/issues/197